refactor: move util.ts, types.ts, model-enhancer.ts to lib/
Moves 3 files from root to lib/ directory:
- util.ts (12 import updates)
- types.ts (4 import updates)
- model-enhancer.ts (0 import updates)

Root folder reduced from 9 to 6 .ts files.
All imports updated to reflect new paths.
Tests pass (pre-existing mock issues remain).
