+++ providers/cline-models.ts
@@ -10,8 +10,8 @@ import {
 	DEFAULT_FETCH_TIMEOUT_MS,
 	DEFAULT_MIN_SIZE_B,
 } from "../constants.ts";
-import type { ProviderModelConfig } from "../types.ts";
-import { cleanModelName, fetchWithRetry, isUsableModel } from "../util.ts";
+import type { ProviderModelConfig } from "../lib/types.ts";
+import { cleanModelName, fetchWithRetry, isUsableModel } from "../lib/util.ts";
 
 interface OpenRouterRaw {
 	id: string;
