+++ providers/kilo.ts
@@ -24,7 +24,7 @@ import {
 	createCtxReRegister,
 } from "../provider-helper.ts";
 import { registerUsageWidget } from "../usage/widget.ts";
-import { cleanModelName, logWarning } from "../util.ts";
+import { cleanModelName, logWarning } from "../lib/util.ts";
 import { loginKilo, refreshKiloToken } from "./kilo-auth.ts";
 import { fetchKiloModels, KILO_GATEWAY_BASE } from "./kilo-models.ts";
